+++ custom_components/adaptive_irrigation/const.py
@@ -54,3 +54,13 @@ SCAN_INTERVAL_MINUTES = 15
 TREND_HOURS = 6
 CALIBRATION_FOLLOWUP_SECONDS = 1800
 STALE_SENSOR_HOURS = 8
+
+ENTRY_TYPE_SYSTEM = "system"
+ENTRY_TYPE_ZONE   = "zone"
+
+HOUR_LABELS = [
+    "12:00 AM", "1:00 AM",  "2:00 AM",  "3:00 AM",  "4:00 AM",  "5:00 AM",
+    "6:00 AM",  "7:00 AM",  "8:00 AM",  "9:00 AM",  "10:00 AM", "11:00 AM",
+    "12:00 PM", "1:00 PM",  "2:00 PM",  "3:00 PM",  "4:00 PM",  "5:00 PM",
+    "6:00 PM",  "7:00 PM",  "8:00 PM",  "9:00 PM",  "10:00 PM", "11:00 PM",
+]